requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0

# ML stack
numpy==2.1.2
scikit-learn==1.4.2

# Multipart/form uploads (FastAPI)
python-multipart==0.0.6

# Testing
pytest==7.4.2

# HTTP async client
httpx==0.27.2

# Environment variables
python-dotenv==1.0.1

# LLM client
groq==0.4.2
